move slider-timer to external module
